auth-service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx 
pydantic
python-jose
requests
idna
grpcio
grpcio-tools
pytest